json-config: extend include_path for deps
Currently elp does not find included files from a dependency's
own (and properly configured) include_dirs, since they are not
added to include_path.

It is not always a viable workaround to put those deps into apps.
When a dep contains a parse_transform that is needed by
erlang_service_lint.erl, it must be classified as a dep for
its ebins to be available to erlang_service.