json-config: also put deps include_dirs into include_path (#107)
Summary:
Currently elp does not find included files from a dependency's own (and properly configured) include_dirs, since they are not added to include_path.

As far as I can tell it is not always a viable workaround to put those deps into apps.
When a dep contains a parse_transform that is needed by erlang_service_lint.erl, it must be classified as a dep for its ebins to be available to erlang_service.

Pull Request resolved: https://github.com/WhatsApp/erlang-language-platform/pull/107

Reviewed By: alanz

Differential Revision: D77531812

Pulled By: robertoaloi

fbshipit-source-id: 078f48ed2671492828f9378aca50d4f0b72ce832