Take global includes into account for deps for JSON projects
Summary:
In D77531812 the project loading via JSON files was patched to fix include paths for dependencies. But the patch did not take into account the global includes.

This change normalizes the way apps and deps handled includes, for consistency. Ideally, we should have snapshot tests for the project loading mechanisms, especially for the ones we don't use internally (like the JSON one).

Reviewed By: TD5

Differential Revision: D77540655

fbshipit-source-id: fd7a310efc404c89f4755240b3eb912f981891dc