Make eqwalizer snapshot tests depend on OTP version
Summary:
When adding a pipeline for OTP 28, the eqwalizer snapshot tests started to fail due to some expected changes between the two OTP versions.
To be able to support this scenario, I am making the snapshot tests depending on the current OTP version.
In practice, this means to duplicate the test fixtures from `X.pretty` to `X-OTP-26.pretty` and `X-OTP-27.pretty`. In a subsequent diff, 28 specific fixtures will be added.

The only change worth reviewing in this diff are the ones to the `main.rs` file.

Reviewed By: TD5

Differential Revision: D77536289

fbshipit-source-id: b80905893be4df23c7d91a293a94fa285e0259f4