update tests to remove "--no-buck-bxl"
Summary:
Update the tests to use ELP defaults, so automatically using buck bxl.

Some of the eqwalizer expected test results had to change, as the buck2 project model has a separate app and test target, so the diagnostics arising from the test suite do not appear.

They do stil for the rebar3 case.

Reviewed By: robertoaloi

Differential Revision: D77732211

fbshipit-source-id: 2313dc781c165e03d5e213ad09df5d94342fa114