Remove all usages of `BuckQueryConfig::Original`
Summary:
Remove all usages of `BuckQueryConfig::Original`, and the enum variant as well.

A follow-up diff will merge `BuckQueryConfig` and `BuildGeneratedCode`, which we still want to have available as a fallback.

This also removes the CLI argument `--no-buck-bxl`x, which is no longer used.

It does not remove `--buck-generated`, as this is still used in some places. That cleanup will happen later.

It also removes the config options as expected from the language server for using bxl and generating code.

Reviewed By: robertoaloi

Differential Revision: D77739226

fbshipit-source-id: fa0e8b4fe7e257721b3382216274bede156d6d7c