Fix W0037 unspecific include for buck generated files
Summary:
Our current `elp.bxl` script returns a full path instead of a directory for the `includes` field for generated files.

This breaks the reverse lookup for W0037, so the unspecific_include diagnostic is not reported.

Add a check on the ELP side for this, and remove the file name if present.

Reviewed By: TD5, robertoaloi

Differential Revision: D77792503

fbshipit-source-id: 90d31106caca8828a07e82d82a4a4b93fd4114c4