Keep original buck TargetFullName in AppData
Summary:
We use `AppData` as our general representation of an Erlang application in ELP, originally in the rebar3 sense.

The ELP design standardises on this information, whatever the origin of the project data in an ELP Project.

For a buck2 project, we use the last part of a buck2 target as the `AppData::name`.

So the buck target `cell/path/to/app:app_name` would give rise to "app_name".

In some circumstances involving test fixtures we cannot guarantee that the derived "app_name" is unique.

This diff adds the original full buck target name to `AppData`, so this can be used as needed.  It is used in a later diff, still unpublished.

Reviewed By: TD5

Differential Revision: D78160505

fbshipit-source-id: 3b14d50709328b53079d648c5eb2ae6a86311d6d