Properly clear diagnostics when a file is closed
Summary:
When we process a `DidCloseTextDocument` we sent a `PublishDocuments` notification with empty diagnostics to the server, to clear its diagnostic store.

But if at the time of closing the file we are calculating diagnostics for it, these will be sent when the diagnostic calculation ends.

This diff ensures we clear all current diagnostics (except eqwalizer project ones, if enabled) when we close the file, and does not send diagnostics for unopened files.

Reviewed By: TD5

Differential Revision: D78275536

fbshipit-source-id: f4d380e0022102dcb74151f027df4d5b0cdd58c5