Track original file when resolving includes
Summary:
When `erlc` resolves an `-include_lib(..)` directive it does so with the search path of the original file.

Up to now ELP has been taking a lenient approach to include file resolution, and has not precisely resolved according to the explicit dependencies of a given application. The application is determined from the originating `FileId`.

This diff keeps track of the originating `FileId`, so it can be used for more precise tracking in a later diff

Reviewed By: TD5

Differential Revision: D78165061

fbshipit-source-id: 225c1af47abea00616fc2763a26c6c89ae2db55f