Introduce lazy_function_matches macro to avoid duplication
Summary:
# Context
Diagnostics are calculated for every `file_id` (i.e. module) in the system, causing multiple calls to extract the function matches. To avoid the overhead, the `lazy_static` macro is usually used.
# Problem
The code required to specify matches is quite verbose.
# Solution
Create a helper macro that minimizes the amount of code written by the lint author.

Note: ideally, we'd get rid of the multiple calls in the first place. That could be done as a next step.

Reviewed By: TD5

Differential Revision: D80342732

fbshipit-source-id: 4cd2f8635b12e083feb37c369f0703cace9b7c56