Introduce SsrPatternsLinter trait
Summary:
Similar to the `FunctionCallLinter` introduced in D80341881, this introduces a `SsrPatternsLinter` trait which abstracts away a lot of the implementation details for Ssr-based linters, allowing the ELP developer to focus on the more important callback functions.

We convert the `UnnecessaryFoldToBuildMap` linter to showcase how the linter can be used in practice.

In the process, fix a few minor bugs in the linter, which emerged from the conversion.

Reviewed By: TD5

Differential Revision: D80462420

fbshipit-source-id: c3d047ff6a020e17194f7a93a0df85775cda6260