Do not publish the Windows extension to the Marketplace yet
Summary:
Support for Windows is still in progress. Tests are failing and users report several issues. See https://github.com/WhatsApp/erlang-language-platform/issues/26 for details.
Until we have a stable Windows CI pipeline which runs all tests, do not publish the Windows version of the extension to the official marketplace.
The extension can still be manually downloaded from the ELP website as an artifact and installed as a .VSIX package.

Reviewed By: TD5

Differential Revision: D80931197

fbshipit-source-id: 5e3f1a2c44575464921877d4395d241095dc7829