Convert apply_to to a apply_to_regex
Summary:
This is a best-effort conversion from the deprecated glob to a regex, using a hand-written glob-to-regex compiler.  We could use glob.translate if we were on python 3.13 or we could use fnmatch.translate if it supported '**'.  Both produce less readable results than I think the hand-rolled parser.

I have inspected each result to look for issues and I didn't see any.

Reviewed By: markisaa

Differential Revision: D81066431

fbshipit-source-id: 2b66f2ca1f70b2b76c53aaf9e715b2951abb2bfa