Improve document symbols calculation and visualization
Summary:
When showing document symbols we were always printing each function in MFA format, accompanied by eventual clauses as "children". In the case where a function was made by a single clause, this led to an unnecessary duplication since both the MFA and the full-version (including arguments) was shown.

This diff ensures that, for functions with a single clause, only the MFA format is visualized.

In the process, a few other improvements were applied to the way we calculate document symbols:

* The logic for extracting information such as name and range was moved to the `module_data` for clarity and reuse
* The `to_document_symbol` now returns an `Option`, making the code slightly simpler
* The `document_symbols` module now avoids operating at AST level directly

One slightly annoying thing is that we don't store the function clause definitions as part of a function def, so we need to reach out to the `def_map` to extract them

Reviewed By: TD5

Differential Revision: D81321568

fbshipit-source-id: 180150b0927b7640863c2c38d470fceaf6236bcb