Show suite code lenses even if all/groups cannot be evaluated
Summary:
If the `all/0` and/or `groups/0` function cannot be evaluated, the `ct_info` function returns an error, in which case we don't show any CT code lenses for the module.

This is not necessary, and there's still value in showing the suite-level code lenses (`Run all tests`, `Debug`, `Open REPL`).

Reviewed By: jcpetruzza

Differential Revision: D81325221

fbshipit-source-id: 4e70f0c56426a414ae412f6ddb4f6e4c93a6714d