Be able to override diagnostic severities via config
Summary:
Severity is currently hardcoded by linters and there's no mechanism to override it. This is inconvenient for two reasons:

* Changing the severity for the linter requires an ELP release
* The linting experience cannot be easily customized by the external community

This is a first step towards a more generic linter configuration mechanism for ELP, which will incrementally expanded.
Since this is currently respected only by the linters implementing the `Linter` trait, we keep the feature currently undocumented.

Reviewed By: TD5

Differential Revision: D81443678

fbshipit-source-id: 17783df15cabb0f7c16493a461e457fd214b2bc6