Support custom fixes in FunctionCallLinter trait
Summary:
Allow instances of the `FunctionCallLinter` trait to implement two new custom methods:

* `is_valid_match`
* `fixes`

The first one, given the *call context* allows to validate each match and return a linter-specific context which is passed around to the other functions.

The `fixes` method allows a linter to specify custom fixes. It receives the context produced by the `is_valid_match` method as part of the *match context*.

The `debugging_function` is then converted to use the new approach, becoming much simpler.

The templating mechanism from the `helpers` module is totally gone, and replaced by the new trait-based one.

There's one performance implication: when the `FunctionCallLinter` trait was introduced, we started collecting "specs" and performing a fold only once. While this approach was convenient, it would have made implementing this new functionality a lot more complicated. Instead, I propose we sacrifice the extra folds (after all, that was state of the world before traits) for the time being. This brings the function call linters closer to the SSR-based ones. Then we can think of a different approach for folding/visiting.

Reviewed By: jcpetruzza

Differential Revision: D81516582

fbshipit-source-id: 6b9dd28b91eb272a670c5bf8679660c600c8611a