detect new project source files
Summary:
# Context

When working with a buck2 project, we do not simply use source roots, as a given buck2 target can be configured to use files and dependencies from anywhere.

To cope with this, we build an include-file lookup structure at project loading time, which is partly populated with information from the output of the bxl query.

If a new `.hrl` or `.erl` file is added, it is not obvious which target is should be part of, to update the relevant application and include file lookup table.

# This Diff

This diff takes a conservative approach to files being added, as they should be a reasonable rare occurrence.

It checks for a notification of a newly created source file (once in the `Running` state), and if so triggers a project reload.  The query for this happens in a separate thread, and when there is a valid result it is applied.  Since there is only an additional file, most of salsa should still be valid, so it does not cause cascading changes.  Existing file ids are retained, as Vfs is not flushed in the process.

Reviewed By: robertoaloi

Differential Revision: D81783795

fbshipit-source-id: 19012379179d0d841bc4b0b24ccd3011e1bfab5d