Mitigate false positive undefined function for lager
Summary:
We received various reports of false positives for the `W0017` (`undefined_function`) linter. In both cases this was due to the usage of the [`lager`](https://github.com/erlang-lager/lager) logging library, which relies on a parse transformation, causing functions such as `lager:warning/X` or `lager:error/X` to be recognized as undefined.

While the medium-term plan is to allow linters to get a custom config such as:

```
[linters.undefined_function]
excludes = [
  "lager",
  "some:call",
  "some:other/3"
]
```

For the time being, add `lager` to the hard-coded exclusion list, to mitigate the issue, reducing noise for end users.

Reviewed By: TD5

Differential Revision: D82094926

fbshipit-source-id: e73eb0ca4315a17e63d35938801badd44d7ac7fd