Allow function call linters to specify a list of matches to exclude from processing
Summary:
Expand the `find_call_in_function` capabilities to allow excluding specific functions by using a vector of *FunctionMatch*.
Then, add an optional `excludes_functions` method to the `FunctionCallLinter` trait.
Use the new method to simplify the code in `undefined_function` to specify the exclusion list at a higher level.
This is in preparation of making the list of exclusions configurable via the TOML file, which will happen next.

Reviewed By: TD5

Differential Revision: D82526373

fbshipit-source-id: 1f07c8495655103d479dc76b6837b3968edb7dba