Allow excluding function call matches via config
Summary:
Add an `exclude` config option for linters implementing the `FunctionCallLinter` trait, allowing to exclude matches via config.
Since the implementation uses `lazy_static` behind the scenes, this is only applied after a server restart.

This will later be used to avoid hard-coding the exclusion list for the `undefined_functions` linter.

For the time being, I'm using the default deserialization, so that one can write:

```
[linters.undefined_function]
exclude = [
  {type = "MFA", mfa = "my:function/2"}
]
```

This can be later refined not to require the boilerplate from users, but auto-detect the M, F or MFA.

Reviewed By: alanz

Differential Revision: D82535853

fbshipit-source-id: fe7bd54c0f1e6c803a6bca16b8951e27ac4dcc58