Add new option to mark Erlang Service warnings as errors
Summary:
An Erlang project can be compiled with the [`warnings_as_errors`](https://www.erlang.org/doc/apps/compiler/compile.html#file/2) option, which causes warnings to be treated as errors.

ELP does not respect this option when it comes to diagnostics produced by the Erlang Service (the Erlang sidecar running next to ELP), which can create confusion in the user: if `warnings_as_errors` is specified, ELP would only raise a warning (both in the IDE and in the CLI), but the project would fail compilation.

Here we add a new configuration section for the `[erlang_service]`, where the option can be explicitly passed. Later on, we could automatically expand this mechanism to auto-extract the option from either a rebar3 or a buck2 project. For now, that's considered out of scope.

Since the Erlang Service validates Erlang modules using a fork of the [`erl_lint`](https://github.com/erlang/otp/blob/master/lib/stdlib/src/erl_lint.erl) linter, which does not respect the `warnings_as_errors` option, we tweak the result on ELP side.

We also document the new config section.

Reviewed By: alanz

Differential Revision: D82628908

fbshipit-source-id: f3ced1cf333b463657cbc13113129bb15533a9d6