BE: Use a vec for load telemetry
Summary:
We currently send a summary telemetry message when a progress bar ends, with each of the steps included in it.
This is currently stored in rust as a `FxHashMap` of segment title and time taken. This does not keep the relative ordering of the segments, so detailed knowledge of the originating code is needed to understand the order.

Simplify this by using a `Vec<(name,time)>` instead.

Reviewed By: robertoaloi

Differential Revision: D82953458

fbshipit-source-id: 8f37e3135d5781505c60c3f30556273859eec7f9