BE: Include a formatted start and end time stamp in telemetry
Summary:
The ELP telemetry currently includes a start time, as a struct having seconds since the Unix epoch and nanoseconds.

This gives us an accurate start time, as we cannot rely on the SCUBA logged time, as it is the time the message is received, which is after many unpredictable events through the middleware.

It also includes a duration in milliseconds.

It is hard to work with these in SCUBA, especially for casual comparisons.

This diff introduces a string representation of the start time as a time stamp to millisecond resolution, as well as an end time stamp, computed from the start time and duration.

Reviewed By: robertoaloi

Differential Revision: D82956324

fbshipit-source-id: 8201875d320cca23358bd1c6d780d4c536e1c557