Convert unused_macro linter to use new GenericLinter trait
Summary:
Define a new `GenericLinter` trait for linters which do not use either the `find_call_in_function` helper or SSR patterns.
As an example, convert the `unused_macro` linter to the new trait.

Apart from simplifying the way linters are written, the big advantage is that linters using the trait can leverage common infrastructure (e.g. including an ignore fix) and they can now be configured / enabled via config.

Note: the `None` variant for the `DiagnosticTag` was eliminated in favour of the standard `Option` type, which allowed some code simplifications.

Reviewed By: alanz

Differential Revision: D82638562

fbshipit-source-id: 25c63591f4ae60ea4e89cbbd79e91e9b95d404ed