Split token identification and documentation extraction
Summary:
The API to extract documentation currently takes a *position* as the input and returns an `Option<Doc>`. This is not ideal, since it collapses two cases into one: `None` can be returned if no meaningful token is identified or if no documentation is available for it.

As a first step towards introducing telemetry for hover operations, have a clearer distinction between the two cases.

Here we:

* Split token extraction and fetching documentation
* Introduce a `DocResult` struct which stores additional information about the input token
* Introduce basic telemetry for the *hover* operation

This will expanded in a subsequent diff to include more information about the actual documentation.

Reviewed By: alanz

Differential Revision: D82948615

fbshipit-source-id: c0dbc2bb99519ef5522288ed5bc910f34105bc54