Use buck2 cell info to resolve target directories
Summary:
## Context

We have been assuming that the buck2 targets we care about are  all in the same buck2 cell, corresponding to the buck root.

This is not always the case, and we are starting to use targets from diverse locations.

## This diff

Update the target to absolute directory processing to use the buck2 target cell information for the given project.

Reviewed By: TheGeorge

Differential Revision: D83245484

fbshipit-source-id: 8410ba36992d6369dccf09a85ffc780d40e33500