1/n: Add field to track generation source file inputs for buck2 generated targets
Summary:
## Context

Our generated targets have output in the `buck-out` directory, which is isolated from the normal buck2 build because we use an isolation directory when we run `elp.bxl`. This means we cannot watch these files for changes when doing a buck2 build, and must monitor the inputs from the targets instead.

D83155666 extends `elp.bxl` to provide this information.

## This diff

Capture this information in the buck project model when running `elp.bxl`.  This sets the stage for using it higher up in the stack.

Reviewed By: robertoaloi

Differential Revision: D83255606

fbshipit-source-id: 9cf4cbee34a1f5ec0a0a5da7ace5d95de0de8adf