3/n: Reload project if a buck2 generated input changes
Summary:
## Context

We are now watching for changes to the inputs of targets that are generated by buck2.

But our existing logic to detect if a project needs reloading is based on hierarchical containment of the changed file inside the project somewhere.

For generated targets this is not always the case, so we need some additional logic.

# This diff

We now keep a map for each generated target input to the application and this project it belongs to, via an `AppDataId`.

This allows us to replace the watched path with one inside a project, and use the normal project reload determination logic.

Reviewed By: TheGeorge

Differential Revision: D83257306

fbshipit-source-id: 1e251d594e37c8c7edc08859722614b66a7f5ba3