Fix flaky tests using buck cell info
Summary:
We now have buck2 cell information, which is cached in a mutex variable as it should not change for the life of a project.

But in tests we set it to fixed values, and check the result.  When tests are run in parallel, this can cause flakiness as values from one test are seen in another.

Add a global mutex around the test setup to prevent this.

Reviewed By: robertoaloi

Differential Revision: D83479480

fbshipit-source-id: efdeaf3c1f7f2c7b3cee6ec5f0ac0e4e90edcece