Include some of the 'erlang' functions when searching for dynamic calls
Summary:
Extend detection for dynamic calls to the most common ones from the `erlang` module. Given that some of these calls refer to functions by arity, extend the original mechanism to support both variants.

Later in the stack we will use a more performant data structure and a macro to reduce verbosity.

Reviewed By: jcpetruzza

Differential Revision: D83840802

fbshipit-source-id: 1fc91f1b368cda4fb3b12ee5fb934871ee672cad