Include dynamic calls when finding references
Summary:
In D41220041 (Nov 2022), we started reporting only direct references for the "find references" functionality. The idea was to filter out references such as the entry in the export list, which were adding noise in the UI, making it harder to find actual references.

Dynamic calls were incorrectly classified as "other", causing dynamic calls not be returned as references (but still working in rename operations, since they applied to *all* references).

Reviewed By: alanz

Differential Revision: D83967812

fbshipit-source-id: bd89c247e89f5823f542c8598b7a28d05b87c9b5