Update loader progress on percent change only
Summary:
A project can have thousands of files. We report loader progress as the VFS loader loads these, presented on the VS Code status bar as a percentage figure.

This means that we send multiple duplicate messages while loading, as the percentage number changes relatively slowly compared to the number of files loaded.

This onslaught of messages can however overload the LSP client, resulting in a status bar that does not accurately reflect the current ELP loading state as it updates from the backlog in a received queue of messages.

## This diff

This diff reduces the number of messages sent, without losing any information, by keeping track of the current percentage value, and only sending a progress message when it changes.

Reviewed By: TheGeorge

Differential Revision: D84140378

fbshipit-source-id: 13fd4e39b87010c70b692abd7aeb15beddef2dbd