server: only register dynamic capabilities once
Summary:
When running in LSP mode, we do not immediately advertise all the ELP cababilities to client, holding back outline support.

This is because while we are doing our initial loading we do not process request messages from the client, and the VS Code one only asks for the outline once, when a file is first opened.

So we register dynamic capabilities once, when ELP switches to `Running` state after completing the initial file loading.

The current implementation registers the dynamic capabilities when loading is complete, which means that if the project model changes and we reload it will register again. This results in a fresh outline request, and the VS Code client then presents two outlines for the same file.

## This diff

So in this diff we add a flag in the server state and only register the dynamic capabilities once.

Reviewed By: robertoaloi

Differential Revision: D84244032

fbshipit-source-id: b1b9b2a5ddd935a58659ff13b92407a265cee34a