Point Cargo.toml to tree-sitter-erlang crate
Summary:
The conditional in `Cargo.toml` is problematic, since it produces a different lock file if compilation happens on GitHub via Cargo. Apart from polluting the workspace with a change, it also makes it hard to use the `--locked` option when building (e.g. part of the standard cargo options in Homebrew).

Simply point to crate and leave a comment explaining how to modify the `Cargo.toml` locally for easier development.

Reviewed By: alanz

Differential Revision: D84346947

fbshipit-source-id: 25fe550e0feec7c2a2567debfaecf8318a0c523a