Introduce simple_snapshot_output_contains helper, fix failing testcase
Summary:
The testcase was not using the `simple_*` helpers, causing it to run with a buck2 config on GitHub. Buck2 is not yet configured there, causing the test to fail.

Introduce a new helper that a testcase to match on the output, rather than having a completely deterministic output (memory can vary). By using the helper, apply the same filtering on the buck feature as in the other tests.

Reviewed By: alanz

Differential Revision: D84498544

fbshipit-source-id: 6cd58f05914cfa218496280c0e4331c2f162980a