Allow overriding is_enabled property via config for trait linters
Summary:
# Context

For linters based on [traits](https://whatsapp.github.io/erlang-language-platform/docs/contributing/linters/linter-traits/) we can customize their behaviour via config. Here we extend the configuration mechanism we already use for `severity` and other properties to the `is_enabled` flag. This mechanism will eventually supersede the `enabled_lints` and `disabled_lints` [configuration properties](https://whatsapp.github.io/erlang-language-platform/docs/get-started/configure-project/elp-lint-toml/#enabled_lints).

# Why having a new way to enable and disable linters?

Currently, the `.elp_lint.toml` allows users to specify which linters are enabled or disabled. But the configuration file works in a very counter-intuitive way: say a linter is "disabled by default". To enable it, a user cannot simply add it to the `enabled_list`, since that would cause **only that linter** to run. A user would need to add all available linters to the config, which is clearly not a maintainable solution.

By supporting an override mechanism, it will be possible to - for example - have the following configuration:

```
[linters.my_linter]
enabled = true
```

Which would keep the *enabled* status for all other linters to their default value, ensuring `my_linter` is enabled.

# This diff

This diff introduces the new overriding mechanism. We still keep it undocumented, until all linters are migrated to use a trait.

Reviewed By: TheGeorge

Differential Revision: D85055745

fbshipit-source-id: f84d203c380a45dcbafae5d09dc9141d42a12e91