Allow filtering of OTP apps in indexing
Summary:
Many applications are included in the Erlang/OTP release, but they are not used by a project. Processing them leads to unnecessary work and memory consumption. This diff adds a new `otp` configuration section to the `.elp.toml` file, allowing users to customize a list of OTP applications to exclude from the analysis.

* Add new `OTP` section to ELP Config
* Pass reference to the whole ELP configuration to the project loading function
* Forward OTP config section to OTP discovery function
* Perform filtering
* Add documentation on how to use the new feature

Reviewed By: TheGeorge

Differential Revision: D83974071

fbshipit-source-id: eae11cd06eca9380d00ca5f7a3a70bb659d6ab71