Fix incorrect project load on go to def in thrift file
Summary:
For a buck project having code generation, the `elp.bxl` script returns the various generated source paths in the `buck-out` directory. But the paths it returns include symlinks.

VS Code resolves symlinks when a file is opened, and all the communication with ELP as an LSP server is done using the canonical (no-symlinks) path.

When we go to a definition, we provide a path that may have a symlink.

We recently landed a diff titled `[ELP] check if a file is in a project when it opens` which checks if a file outside the project root in fact belongs to a project, via the list of `applicable_files` for a buck project.

For a symlinked file opened in VS Code and returned as canonical, this lookup fails.  The failure also shows up another bug, in that the check to see if the file belongs to a project uses function `file_app_data` which falls back to checking for a `SourceRoot`, and panics for a file outside the root.

## This diff

This diff fixes both problems.
- It canonicalizes (without symlinks) the `applicable_files` for buck projects
- It rather uses `app_data_id_by_file` which does a simple loookup to check if the file belongs to a known project.

Reviewed By: jcpetruzza

Differential Revision: D85429234

fbshipit-source-id: e1f4867f8c66e69e219defd3ea742f431b994a21