Specify SSR match as Ad-Hoc diagnostic
Summary:
Structured Search (and not yet Replace) SSR is a way of specifying a "semantic grep", which searches by comparing the AST rather than text. We use it as a building block for diagnostics.

But often we want to just do a quick search for a pattern in a file or code base.

This diff allows us to do this, by specifying it in a config file.

Reviewed By: TD5

Differential Revision: D85340131

fbshipit-source-id: f1e4ce515a535cc6032483dc39297eab4e6eecd5