Add @OTPVersionDependent tag support for tests
Summary:
**The core change is in `main.rs`**

There is no need for most snapshots to be generated for 3 different OTP versions. Only `custom.erl` needs it.
Add support for tag `OTPVersionDependent`, similarly to `generated`, to toggle this behaviour.

Add tag to `custom.erl` (on an existing line so as to not change the lines of error messages).

Reviewed By: michalmuskala

Differential Revision: D85602616

fbshipit-source-id: 4c85ecb3e2337d60cf83d13b7449e1b20246d9ed