1/n: buck targets; Add --buck-quick-start CLI argument
Summary:
## Context

ELP with buck2 via elp.bxl has a slow startup time.  This time is driven primarily by how long it takes the buck2 daemon to simply load the various BUCK files.

Given that the buck targets may generate code, which is triggered in elp.bxl, this is unavoidable for full-fidelity project loading.

There is a much quicker command, `buck2 targets`, which does not do any build steps for code generation, but does run all the buck2 macros.  This gives us an accurate project representation for that portion of the project that is not generated.

Our ELP internal project model has the same representation for the initial subset query as for the full one, so we can do the quick query followed by the slow one without having to discard the work already done.  In particular, it means simply adding source files to existing targets.

## This diff stack

This diff stack brings in this two step process, so that in IDE mode we first do a quick query and start file loading. As soon as the initial file loading is complete, we can go operational.
At the same time as we start file loading, we initiate the slower `elp.bxl` query which also triggers code generation.  When that completes, we update the internal model and load the missing files.

The result is a quicker initial operation, albeit without access to any generated files, which are loaded when ready.

## This diff

This diff makes a start, by bringing in a command line argument to select the new behaviour. This defaults to off.

Reviewed By: TheGeorge

Differential Revision: D83660182

fbshipit-source-id: 9aea01d68fbacde5fb5fc032a6adbb522207741f