6/n: buck targets: dump target list for comparison
Summary:
Add a test facility to `elp project-info` to be able to compare the target set arising from the `buck2 targets` call vs using `elp.bxl`.

This explicitly excludes any targets marked as generated by `elp.bxl`, as these will not appear in the bare `buck targets`.

It sorts the output according to whether it is an `App` or `Dep`, the app name and then the original buck target name, for investigation and confirmation.

Reviewed By: TheGeorge

Differential Revision: D83829236

fbshipit-source-id: be27273fdaf6d095b59eccda0cf74165510fe0b2