3/n: buck targets; Improve CLI project-info output
Summary:
We want to be able to compare the targets generated via `buck targets` with the ones generated via `elp.bxl`.

Each of these emits the targets in their own essentially random order.

This diff tweaks the processing of the `elp project-info` command so we can more easily compare the results with and without the new `--buck-quick-start` flag.

Reviewed By: robertoaloi

Differential Revision: D83662426

fbshipit-source-id: df15ba80a397f9b7b0614d045cb1d53c2f98f66a