4/n: buck targets: explicitly check for the prelude as a target, for origin
Summary:
A buck target is classified as an `App` or a `Dep`, according to whether it is part of the project being loaded or is one of its dependencies.

The `elp.bxl` script [provides this information](https://github.com/facebook/buck2/blob/db5bf933afd23ac3306b7f3627819046b6973534/prelude/erlang/elp.bxl#L94-L96) by filtering the included targets for erlang_app/erlang_dep only, and marking these as apps.

It then generates the full configured graph from this set of targets, and the additional targets (erlang_apps only, no tests) are marked as deps.  This is the step that takes a long time, and which we are avoiding in the initial step.

To achieve the same for the first step we also do two `buck targets` queries, one for the included targets as used in `elp.bxl`, the second for the buck config `deps_targets`, which are not used in `elp.bxl` as they are computed from the dependencies in the configured graph for the included targets.

Reviewed By: robertoaloi

Differential Revision: D83758970

fbshipit-source-id: ea84c51e76a63a1418cbbdd72718bdca076f4808