8/n: buck targets: Integrate into server mode
Summary:
This diff stack is about improving the initial IDE start experience when using a buck project that uses buck rules to generate some files, and for which that step takes a long time.

It does so by initially using a much faster buck command that does not do any code generation, but does expand all buck macros, `buck targets`.  Based on this result, configures the ELP project model and goes operational with the given project file set.

While loading these files, it also does the normal full `elp.bxl` query to generate the additional required files.  Once this is done, the project and its files are reloaded. This is basically a no-op except for adding in the newly generated files.

## This diff

This diff makes the changes to ELP in server mode to enact the process outlined above.

Reviewed By: robertoaloi

Differential Revision: D84343272

fbshipit-source-id: 8944767241f6f2ecb6c076e8c11a1a8c80057570