Tweak telemetry manager test to not be flaky
Summary:
The `test_lifecycle` test is intended to confirm that we do get multiple segments.

On a lightly loaded machine, the timing should be zero. This is not always the case, so to avoid flakiness we manually zero out the timing, so the essence of the test is not flaky

Reviewed By: jcpetruzza

Differential Revision: D85661410

fbshipit-source-id: b7d1130929692360a1284c78787f89f4e537cd88