Add `elp ssr` command
Summary:
We have the ability to specifiy an AdHoc diagnostic in a config file, from the diff titled `[ELP] Specify SSR match as Ad-Hoc diagnostic`.

We now simplify usage by allowing it to be given on the command line too, so a simple

```
elp ssr "{_A, _B}"
```

will find and report all 2 element tuples in the project.

Note: the full syntax for ssr can also be used if guards are required, e.g.

```
elp ssr "ssr: {_X = _Y} when _X == foo."
```

Reviewed By: michalmuskala

Differential Revision: D85443930

fbshipit-source-id: 108e2a583437fb3a6e702c21aef874395b376e7b